requirements.txt
# Core Dependencies
asyncio>=3.4.3
aiohttp>=3.9.0
aioredis>=2.0.1
python-dotenv>=1.0.0

# Network & Communication
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.25.0

# Data & State Management
redis>=5.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
locust>=2.17.0

# Monitoring & Metrics
prometheus-client>=0.19.0

# Utils
psutil>=5.9.6
colorlog>=6.8.0
